backend/requirements.txt
flask
flask-cors
gunicorn
python-dotenv
openai
transformers
torch